Add support for filtering entries by patterns in profiles
This commit adds two new configuration options, 'require_patterns' and 'exclude_patterns' to profiles. When values are specified in 'require_patterns', entries must contain at least one of the specified patterns in their 'author' or 'title' elements in order to trigger a notification. When values are specified in 'exclude_patterns', entries that contain any of the specified patterns in their 'author' or 'title' elements will be excluded from notifications.